testing moo lexer
