+++ package.json
@@ -13,14 +13,16 @@
     "gaze-cli": "^0.2.0",
     "jest": "^20.0.4",
     "lodash": "^4.17.4",
+    "moo": "^0.3.3",
     "nearley": "^2.10.1",
+    "nearley-moo": "juliankrispel/nearley-moo#patch-1",
     "ramda": "^0.24.1",
     "regenerator-runtime": "^0.10.5"
   },
   "scripts": {
-    "compile": "nearleyc src/bishbosh.ne -o src/parser.js",
+    "compile": "nearleyc src/grammar.ne -o src/compiled.js",
     "test": "yarn compile && jest",
-    "test:watch": "gaze \"yarn test\" \"src/**/[!parser]*.*\""
+    "test:watch": "gaze \"yarn test\" \"src/**/[!compiled]*.*\"",
+    "dia": "nearley-railroad src/bishbosh.ne -o src/grammar.html"
   }
-
 }
