+++ src/__tests__/parser.spec.js
@@ -1,14 +1,20 @@
 // The compiler
-import nearley from 'nearley';
 import parser from '../parser';
 
 let p;
 
-beforeAll(() => {
-  p = new nearley.Parser(parser.ParserRules, parser.ParserStart);
-})
+beforeEach(() => {
+  p = parser();
+});
 
 describe('parser', () => {
-  it('', () => {
+  it('parses', () => {
+    // console.log(JSON.stringify(p.feed('aber > Hello there').results, null, 2));
+    // console.log('a, aber, aberdoch -> Hello there my friend');
+    // console.log(p.feed('ab, aber > Hello'));
+    // p.feed('a')
+
+    p.feed('a, aber, aberdoch > Hello world');
+    console.log(p.results);
   });
 });
