--- src/bishbosh.ne
@@ -1,3 +1,0 @@
-@builtin "number.ne"
-@builtin "whitespace.ne"
-
