+++ src/grammar.html
@@ -1,0 +1,196 @@
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<style type="text/css">
+svg.railroad-diagram {
+    background-color: hsl(30,20%,95%);
+}
+svg.railroad-diagram path {
+    stroke-width: 3;
+    stroke: black;
+    fill: rgba(0,0,0,0);
+}
+svg.railroad-diagram text {
+    font: bold 14px monospace;
+    text-anchor: middle;
+}
+svg.railroad-diagram text.label {
+    text-anchor: start;
+}
+svg.railroad-diagram text.comment {
+    font: italic 12px monospace;
+}
+svg.railroad-diagram rect {
+    stroke-width: 3;
+    stroke: black;
+    fill: hsl(120,100%,90%);
+}
+
+</style>
+</head>
+<body>
+<h1><code>command_list</code></h1>
+<div>
+<svg class="railroad-diagram" width="317" height="71" viewBox="0 0 317 71">
+<g transform="translate(.5 .5)">
+<path d="M 20 21 v 20 m 10 -20 v 20 m -10 -10 h 20.5"></path>
+<g>
+<path d="M40 31h0"></path>
+<path d="M276 31h0"></path>
+<path d="M40 31h20"></path>
+<g>
+<path d="M60 31h0"></path>
+<path d="M256 31h0"></path>
+<path d="M60 31h10"></path>
+<g>
+<path d="M70 31h0"></path>
+<path d="M246 31h0"></path>
+<path d="M70 31h10"></path>
+<g>
+<path d="M80 31h0"></path>
+<path d="M236 31h0"></path>
+<rect x="80" y="20" width="156" height="22"></rect>
+<text x="158" y="35">command&#95;list&#95;item</text>
+</g>
+<path d="M236 31h10"></path>
+<path d="M80 31a10 10 0 0 0 -10 10v0a10 10 0 0 0 10 10"></path>
+<g>
+<path d="M80 51h156"></path>
+</g>
+<path d="M236 51a10 10 0 0 0 10 -10v0a10 10 0 0 0 -10 -10"></path>
+</g>
+<path d="M246 31h10"></path>
+</g>
+<path d="M256 31h20"></path>
+</g>
+<path d="M 276 31 h 20 m -10 -10 v 20 m 10 -20 v 20"></path>
+</g>
+</svg>
+
+</div>
+<h1><code>command_list_item</code></h1>
+<div>
+<svg class="railroad-diagram" width="345" height="92" viewBox="0 0 345 92">
+<g transform="translate(.5 .5)">
+<path d="M 20 21 v 20 m 10 -20 v 20 m -10 -10 h 20.5"></path>
+<g>
+<path d="M40 31h0"></path>
+<path d="M304 31h0"></path>
+<path d="M40 31h20"></path>
+<g>
+<path d="M60 31h0"></path>
+<path d="M284 31h0"></path>
+<g>
+<path d="M60 31h0"></path>
+<path d="M284 31h0"></path>
+<path d="M60 31h20"></path>
+<g>
+<path d="M80 31h8"></path>
+<path d="M256 31h8"></path>
+<path d="M88 31h10"></path>
+<g>
+<path d="M98 31h0"></path>
+<path d="M126 31h0"></path>
+<rect x="98" y="20" width="28" height="22"></rect>
+<text x="112" y="35">&#95;</text>
+</g>
+<path d="M126 31h10"></path>
+<path d="M136 31h10"></path>
+<g>
+<path d="M146 31h0"></path>
+<path d="M174 31h0"></path>
+<rect x="146" y="20" width="28" height="22"></rect>
+<text x="160" y="35">&#95;</text>
+</g>
+<path d="M174 31h10"></path>
+<path d="M184 31h10"></path>
+<g>
+<path d="M194 31h0"></path>
+<path d="M246 31h0"></path>
+<rect x="194" y="20" width="52" height="22"></rect>
+<text x="220" y="35">word</text>
+</g>
+<path d="M246 31h10"></path>
+</g>
+<path d="M264 31h20"></path>
+<path d="M60 31a10 10 0 0 1 10 10v10a10 10 0 0 0 10 10"></path>
+<g>
+<path d="M80 61h0"></path>
+<path d="M264 61h0"></path>
+<path d="M80 61h10"></path>
+<g>
+<path d="M90 61h0"></path>
+<path d="M134 61h0"></path>
+<rect x="90" y="50" width="44" height="22" rx="10" ry="10"></rect>
+<text x="112" y="65">","</text>
+</g>
+<path d="M134 61h10"></path>
+<path d="M144 61h10"></path>
+<g>
+<path d="M154 61h0"></path>
+<path d="M182 61h0"></path>
+<rect x="154" y="50" width="28" height="22"></rect>
+<text x="168" y="65">&#95;</text>
+</g>
+<path d="M182 61h10"></path>
+<path d="M192 61h10"></path>
+<g>
+<path d="M202 61h0"></path>
+<path d="M254 61h0"></path>
+<rect x="202" y="50" width="52" height="22"></rect>
+<text x="228" y="65">word</text>
+</g>
+<path d="M254 61h10"></path>
+</g>
+<path d="M264 61a10 10 0 0 0 10 -10v-10a10 10 0 0 1 10 -10"></path>
+</g>
+</g>
+<path d="M284 31h20"></path>
+</g>
+<path d="M 304 31 h 20 m -10 -10 v 20 m 10 -20 v 20"></path>
+</g>
+</svg>
+
+</div>
+<h1><code>word</code></h1>
+<div>
+<svg class="railroad-diagram" width="309" height="71" viewBox="0 0 309 71">
+<g transform="translate(.5 .5)">
+<path d="M 20 21 v 20 m 10 -20 v 20 m -10 -10 h 20.5"></path>
+<g>
+<path d="M40 31h0"></path>
+<path d="M268 31h0"></path>
+<path d="M40 31h20"></path>
+<g>
+<path d="M60 31h0"></path>
+<path d="M248 31h0"></path>
+<path d="M60 31h10"></path>
+<g>
+<path d="M70 31h0"></path>
+<path d="M238 31h0"></path>
+<path d="M70 31h10"></path>
+<g>
+<path d="M80 31h0"></path>
+<path d="M228 31h0"></path>
+<rect x="80" y="20" width="148" height="22" rx="10" ry="10"></rect>
+<text x="154" y="35">/&#91;a-zA-Z&#95;0-9\-&#93;/</text>
+</g>
+<path d="M228 31h10"></path>
+<path d="M80 31a10 10 0 0 0 -10 10v0a10 10 0 0 0 10 10"></path>
+<g>
+<path d="M80 51h148"></path>
+</g>
+<path d="M228 51a10 10 0 0 0 10 -10v0a10 10 0 0 0 -10 -10"></path>
+</g>
+<path d="M238 31h10"></path>
+</g>
+<path d="M248 31h20"></path>
+</g>
+<path d="M 268 31 h 20 m -10 -10 v 20 m 10 -20 v 20"></path>
+</g>
+</svg>
+
+</div>
+</body>
+</html>
